# Improve performance by optimizing glyph lookup