--- a/include/vrv/resources.h
+++ b/include/vrv/resources.h
@@ -65,7 +65,7 @@ public:
     /** Load all music fonts available in the resource directory */
     bool LoadAll();
     /** Set the fallback font (Leipzig or Bravura) when some glyphs are missing in the current font */
-    bool SetFallback(const std::string &fontName);
+    void SetFallbackFont(const std::string &fontName);
     /** Get the fallback font name */
     std::string GetFallbackFont() const { return m_defaultFontName; }
     /** Init the text font (bounding boxes and ASCII only) */