This PR improves the total runtime by optimizing the glyph lookup in resources. Some runtime data with a few pieces of different size:

| Number of pages | Runtime before in s | Runtime after in s | Improvement in % |
| ----- | ----- | ----- | ----- |
| 3 | 0.306 | 0.260 | 15 |
| 10 | 1.732 | 1.525 | 12 |
| 97 | 4.339 | 3.963 | 9 |

Note that after adding the possibility to use several fonts simultaneously, glyph lookup now involves two maps: one for loaded fonts and one for the glyph itself. Changes in this PR:
- We optimize the number of map lookups, mostly by replacing `map::contains` followed by `map::at` with a single `map::find`.
- We cache the glyph that was looked up last. This is helpful, since often the same glyph is requested several times consecutively, e.g. from calls of `GetGlyphWidth` and `GetGlyphHeight`.
- There are a few other (minor) code improvements in `Resources`.